refactor(tmp-toolkit-helpers): DiffFormatter class
